remove capstone, remove adisassembler
