+++ UnitTests/windows/TestDisassembler.cpp
@@ -2,7 +2,6 @@
 // Created by steve on 3/22/17.
 //
 #include "Catch.hpp"
-#include "polyhook2/CapstoneDisassembler.hpp"
 #include "polyhook2/ZydisDisassembler.hpp"
 #include "polyhook2/Tests/StackCanary.hpp"
 #include "polyhook2/MemAccessor.hpp"
@@ -133,7 +132,7 @@ TEST_CASE("Test Instruction UUID generator", "[Instruction],[UID]") {
 	}
 }
 
-TEMPLATE_TEST_CASE("Test Disassemblers x64", "[ADisassembler],[CapstoneDisassembler],[ZydisDisassembler]", PLH::CapstoneDisassembler, PLH::ZydisDisassembler) {
+TEMPLATE_TEST_CASE("Test Disassemblers x64", "[ZydisDisassembler]", PLH::ZydisDisassembler) {
 	PLH::StackCanary canaryg;
 	TestType disasm(PLH::Mode::x64);
 	auto                      Instructions = disasm.disassemble((uint64_t)&x64ASM.front(), (uint64_t)&x64ASM.front(),
@@ -244,7 +243,7 @@ TEMPLATE_TEST_CASE("Test Disassemblers x64", "[ADisassembler],[CapstoneDisassemb
 	}
 }
 
-TEMPLATE_TEST_CASE("Test Disassemblers x86 FF25", "[ADisassembler],[CapstoneDisassembler],[ZydisDisassembler]", PLH::CapstoneDisassembler, PLH::ZydisDisassembler) {
+TEMPLATE_TEST_CASE("Test Disassemblers x86 FF25", "[ZydisDisassembler]", PLH::ZydisDisassembler) {
 #ifdef POLYHOOK2_ARCH_X64
 	// this test is not suitable for x64 due to ff 25 not being re-written
 	return;
@@ -275,7 +274,7 @@ TEMPLATE_TEST_CASE("Test Disassemblers x86 FF25", "[ADisassembler],[CapstoneDisa
 	REQUIRE(Instructions.at(0).hasDisplacement());
 }
 
-TEMPLATE_TEST_CASE("Test Disassemblers x86", "[ADisassembler],[CapstoneDisassembler],[ZydisDisassembler]", PLH::CapstoneDisassembler, PLH::ZydisDisassembler) {
+TEMPLATE_TEST_CASE("Test Disassemblers x86", "[ZydisDisassembler]", PLH::ZydisDisassembler) {
 	PLH::StackCanary canaryg;
 	TestType disasm(PLH::Mode::x86);
 	auto                      Instructions = disasm.disassemble((uint64_t)&x86ASM.front(), (uint64_t)&x86ASM.front(),
@@ -380,7 +379,7 @@ TEMPLATE_TEST_CASE("Test Disassemblers x86", "[ADisassembler],[CapstoneDisassemb
 	}
 }
 
-TEMPLATE_TEST_CASE("Test Disassemblers x64 Two", "[ADisassembler],[CapstoneDisassembler],[ZydisDisassembler]", PLH::CapstoneDisassembler, PLH::ZydisDisassembler) {
+TEMPLATE_TEST_CASE("Test Disassemblers x64 Two", "[ZydisDisassembler]", PLH::ZydisDisassembler) {
 	PLH::StackCanary canaryg;
 	TestType disasm(PLH::Mode::x64);
 	PLH::insts_t Instructions = disasm.disassemble((uint64_t)&x64ASM2.front(), (uint64_t)&x64ASM2.front(),
@@ -397,7 +396,7 @@ TEMPLATE_TEST_CASE("Test Disassemblers x64 Two", "[ADisassembler],[CapstoneDisas
 	}
 }
 
-TEMPLATE_TEST_CASE("Test Disassemblers NOPS", "[ADisassembler],[CapstoneDisassembler],[ZydisDisassembler]", PLH::CapstoneDisassembler, PLH::ZydisDisassembler) {
+TEMPLATE_TEST_CASE("Test Disassemblers NOPS", "[ZydisDisassembler]",PLH::ZydisDisassembler) {
 	PLH::StackCanary canaryg;
 	TestType disasm(PLH::Mode::x64);
 	PLH::insts_t Instructions = disasm.disassemble((uint64_t)&x86x64Nops.front(), (uint64_t)&x86x64Nops.front(),
@@ -418,42 +417,6 @@ TEMPLATE_TEST_CASE("Test Disassemblers NOPS", "[ADisassembler],[CapstoneDisassem
 		for (auto& ins : Instructionsx86) {
 			REQUIRE(ins.getMnemonic() == "nop");
 			REQUIRE(TestType::isPadBytes(ins));			
-		}
-	}
-}
-
-// unreachable code
-#pragma warning(disable: 4702)
-TEST_CASE("Compare x86 Decompilers", "[ADisassembler],[ZydisDisassembler][CapstoneDisassembler]") {
-	PLH::StackCanary canaryg;
-	// Use capstone as reference
-	PLH::CapstoneDisassembler disasmRef(PLH::Mode::x86);
-	auto                      InstructionsRef = disasmRef.disassemble((uint64_t)&x86ASM.front(), (uint64_t)&x86ASM.front(),
-		(uint64_t)&x86ASM.front() + x86ASM.size(), PLH::MemAccessor());
-
-	PLH::ZydisDisassembler disasm(PLH::Mode::x86);
-	auto                      Instructions = disasm.disassemble((uint64_t)&x86ASM.front(), (uint64_t)&x86ASM.front(),
-		(uint64_t)&x86ASM.front() + x86ASM.size(), PLH::MemAccessor());
-
-	SECTION("Check Integrity") {
-		PLH::StackCanary canary;
-		REQUIRE(Instructions.size() == 7);
-		std::cout << Instructions << std::endl;
-
-		for (const auto &p : disasm.getBranchMap()) {
-			std::cout << std::hex << "dest: " << p.first << " -> " << std::dec << p.second << std::endl;
-		}
-
-		for (size_t i = 0; i < Instructions.size(); i++) {
-			INFO("Index: " << i << " Mnemonic:"
-			<< Instructions[i].getMnemonic());
-
-			REQUIRE(filterJXX(Instructions[i].getMnemonic()) == filterJXX(InstructionsRef[i].getMnemonic()));
-			REQUIRE(Instructions[i].size() == InstructionsRef[i].size());
-			REQUIRE(Instructions[i].isBranching() == InstructionsRef[i].isBranching());
-
-			REQUIRE(Instructions[i].getAddress() == InstructionsRef[i].getAddress());
-			REQUIRE(Instructions[i].getDestination() == InstructionsRef[i].getDestination());
 		}
 	}
 }
