+++ polyhook2/Detour/x86Detour.hpp
@@ -9,7 +9,7 @@
 #include "polyhook2/Detour/ADetour.hpp"
 #include "polyhook2/Enums.hpp"
 #include "polyhook2/Instruction.hpp"
-#include "polyhook2/ADisassembler.hpp"
+#include "polyhook2/ZydisDisassembler.hpp"
 #include "polyhook2/ErrorLog.hpp"
 #include "polyhook2/MemProtector.hpp"
 
@@ -19,9 +19,9 @@ namespace PLH {
 
 class x86Detour : public Detour {
 public:
-	x86Detour(const uint64_t fnAddress, const uint64_t fnCallback, uint64_t* userTrampVar, PLH::ADisassembler& dis, const uint8_t maxDepth = c_maxDepth);
+	x86Detour(const uint64_t fnAddress, const uint64_t fnCallback, uint64_t* userTrampVar, PLH::ZydisDisassembler& dis, const uint8_t maxDepth = c_maxDepth);
 
-	x86Detour(const char* fnAddress, const char* fnCallback, uint64_t* userTrampVar, PLH::ADisassembler& dis, const uint8_t maxDepth = c_maxDepth);
+	x86Detour(const char* fnAddress, const char* fnCallback, uint64_t* userTrampVar, PLH::ZydisDisassembler& dis, const uint8_t maxDepth = c_maxDepth);
 	virtual ~x86Detour() = default;
 	virtual bool hook() override;
